requirements.txt
asttokens==2.4.1
attrs==23.1.0
branca==0.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
folium==0.15.0
fonttools==4.44.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
nbformat==5.9.2
nest-asyncio==1.5.8
numpy==1.25.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
patsy==0.5.3
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
prompt-toolkit==3.0.40
psutil==5.9.6
psycopg2==2.9.6
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
SQLAlchemy==2.0.17
squarify==0.4.3
stack-data==0.6.3
statsmodels==0.14.0
tenacity==8.2.3
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
widgetsnbextension==4.0.9
zipp==3.17.0
